CORE-2534 Added remote search to BaseCatalogMainPanel.
Refactored the ToolBar in BaseCatalogMainPanel into a CatalogMainToolBar class.
Added a ComboBox to the CatalogMainToolBar, with remote App search called by an AppTemplateServiceFacade given in its constructor.
Added a protected AppTemplateServiceFacade to BaseCatalogMainPanel, given in its constructor.
Added group ID and name fields to the Analysis model, which are included in the App search results.