CORE-2534 Refactored BaseCatalogMainPanel::selectTool.
Updated selectTool to always set appIdToSelect to the given ID, but only attempt to select the app if it is found in the grid's store first.
This will allow the ID to be set by a search-select-event even when the grid is already loaded with apps in another category, then the event can select the correct category which will then load this app in the grid, then finally the app can be selected by a selectToolNow call in the seed method.
Updated selectToolNow to use the grid store's findModel method.