CORE-2534 Fix for app search selection with dup app names.
Updated the app search ComboBox in CatalogMainToolBar to use a custom store key provider, which is used by a custom ListModelPropertyEditor.
This will allow the ComboBox to find the correct model if 2 apps in different groups have the same name, or even if the same app is listed in different groups, since the combo's SelectionChange event will find the first model that matches the raw text in the ComboBox's text field with its property editor.
Added a lastQueryText to CatalogMainToolBar that caches the user's search query before a selection is made, then resets the ComboBox's raw text field to this value in the Select event listener, so that the custom property editor doesn't display the key provider's key in this field.