CORE-2534 Added App search results sorting.
Added a custom StoreSorter to the search results store in CatalogMainToolBar, so that App names that match the search text are ordered before Apps with only descriptions that match.
The user's query text is now cached before the search service is called with that query, rather than just before the user selects a search result; so that is it available to the custom store sorter.