CORE-3181 Initial refactor of App search ComboBox.
Refactored the App search ComboBox in CatalogMainToolBar into its own class, AppSearchField.
Refactored the App search ComboBox's RpcProxy into its own class, AppSearchRpcProxy.
Added an AppSearchResultLoadEvent and Handler that is fired by the AppSearchField when the user presses enter in, or clears, the search ComboBox.
Updated BaseCatalogMainPanel to listen for an AppSearchResultLoadEvent and populate itself with the search results.
Updated AbstractCatalogCategoryPanel to listen for an AppSearchResultLoadEvent and to clear or restore its selection.
Refactored the deSelectCurrentCategory method into AbstractCatalogCategoryPanel from its child class in the de-webapp repo.