CORE-3181 Added AppSearch cleanup and trigger actions.
Added a trigger button to the AppSearchField that will populate the app grid panel with the search results when clicked, then update its image to an "X" that will clear the search box when clicked, reverting its image to the search icon.
The trigger will also revert to the search icon when a category is selected.
Added cleanup methods to AppSearchField, CatalogMainToolBar, BaseCatalogMainPanel, and AbstractCatalogCategoryPanel that will cleanup their event handlers.