CORE-4163 Refactored AppSearchField for new SearchField.
Updated AppsViewToolbarImpl to build the AppSearchRpcProxy and PagingLoader required by the new AppSearchField.
Reimplemented CORE-3543 functionality from 1.6 in AppsViewPresenter and AppsViewImpl.
Refactored AppSearchResultLoadEvent to take the AppSearchRpcProxy as the source of the event, which the AppsViewPresenter can use to check if it should handle that event.
Refactored AppSearchRpcProxy to use AppListLoadResult and AppsViewToolbarImpl to use AppLoadConfig.
Removed obsolete AppSearch3 classes.